New entity Contrat

+++ src/main/java/tn/esprit/spring/repository/ContratRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.spring.entities.Contrat;
 
-public interface ContratRepository extends CrudRepository<Contrat,Integer> {
+public interface ContratRepository extends CrudRepository<Contrat,Long> {
 }
